player stats component  + subworld component
