+++ client/actions/germanGame.js
@@ -19,14 +19,14 @@
 // }
 
 export const fetchGermanGame = () => {
-  console.log('this is fetch german game')
+  // console.log('this is fetch german game')
   return(dispatch) => {
     $.ajax({
       url: '/api/deutsch/',
       type: 'GET',
       dataType: 'JSON'
     }).done( game => {
-        console.log('fetch german game done', game)
+        // console.log('fetch german game done', game)
         dispatch({ type: 'GERMAN_GAME', game });
     }).fail( data => {
       console.log('fetch german games fail')
