+++ client/components/SubWorld.js
@@ -1,35 +1,20 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { fetchSubThemes } from '../actions/subThemes';
 
 class SubWorld extends React.Component {
 
   constructor(props) {
     super(props);
-
-    this.state = {
-      theme: ''
-    }
   }
 
   componentDidMount(){
-    const search = window.location.search.replace("?theme=", "");
-    this.setState({theme: search})
-    this.props.dispatch(fetchSubThemes(search))
-  }
-
-  renderSubWorlds = () => {
-    const world = this.props.subThemes;
-    return world.map(subtheme => {
-      return(<div key={subtheme.id}>{subtheme.name}</div> )
-    })
   }
 
   render() {
+    const subtheme = this.props.subtheme;
     return (
-      <div id="world_subthemes">
-        <h1>{this.state.theme}</h1>
-        {this.renderSubWorlds()}
+      <div className="world_subtheme">
+        <h1>{subtheme.name}</h1>
       </div>
     )
   }
@@ -37,8 +22,6 @@ class SubWorld extends React.Component {
 
 const mapStateToProps = (state) => {
   return {
-    germanGame: state.germanGame,
-    subThemes: state.subThemes
   }
 }
 export default connect(mapStateToProps)(SubWorld);
